requirements.txt
fastapi
uvicorn
peewee
psycopg-binary
python-dotenv